sdk/tests/conformance/canvas/render-after-resize-test.html

<!--

/*
** Copyright (c) 2017 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->
<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL render after resize test</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"> </script>
</head>
<body>
<div id="description"></div>
<canvas style="width: 100px, height: 100px; border: 1px solid blue;" id="c"></canvas>
<div id="console"></div>
<script>
description("This test ensures WebGL implementations can render correctly after resizing the canvas.");
debug("");

var wtu = WebGLTestUtils;
var gl = wtu.create3DContext("c");
shouldBeTrue("gl != null");

var positionLocation = 0;
var texcoordLocation = 1;
var program = wtu.setupSimpleColorProgram(gl, 0);
var colorLocation = gl.getUniformLocation(program, 'u_color');
gl.uniform4fv(colorLocation, [1.0, 0.0, 0.0, 1.0]);
var buffers = wtu.setupQuad(gl, {
  positionLocation: positionLocation,
  texcoordLocation: texcoordLocation,
  lowerLeftTexCoords: [ 0.0, 0.0 ],
  upperRightTexCoords: [ 1.0, 1.0 ],
  scale: 0.25,
});

const smallWidth = 300;
const smallHeight = 150;
// Changing this size to something smaller produces
// different results. Sometimes wrong, sometimes correct.
const largeWidth = 1200;
const largeHeight = 672;

function render() {
  gl.viewport(0, 0, gl.drawingBufferWidth, gl.drawingBufferHeight);

  gl.enable(gl.DEPTH_TEST);
  gl.clearColor(0,1,0,1);
  gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);

  gl.useProgram(program);
  wtu.drawUnitQuad(gl);
}

function checkForDot(ctx) {
  let w = ctx.drawingBufferWidth;
  let h = ctx.drawingBufferHeight;

  // Round all coordinates to integers.
  const dotWidth  = (w / 4) | 0;
  const dotHeight = (h / 4) | 0;
  const dotX = (w / 2 - dotWidth / 2) | 0;
  const dotY = (h / 2 - dotHeight / 2) | 0;
  const slop = 1;

  wtu.checkCanvasRect(ctx, 0,           0,           w,                   dotY - slop,          [0, 255, 0, 255]);  // top
  let start = dotY + dotHeight + 2 * slop;
  wtu.checkCanvasRect(ctx, 0,           start,       w,                   h - start,            [0, 255, 0, 255]);  // bottom
  wtu.checkCanvasRect(ctx, 0,           dotY,        dotX - slop,         dotHeight,            [0, 255, 0, 255]);  // left
  start = dotX + dotWidth + 2 * slop;
  wtu.checkCanvasRect(ctx, start,       dotY,        w - start,           dotHeight,            [0, 255, 0, 255]);  // right
  wtu.checkCanvasRect(ctx, dotX + slop, dotY + slop, dotWidth - 2 * slop, dotHeight - 2 * slop, [255, 0, 0, 255]);  // dot
}

gl.canvas.width = smallWidth;
gl.canvas.height = smallHeight;
render();
checkForDot(gl); // passes

gl.canvas.width = largeWidth;
gl.canvas.height = largeHeight;
gl.canvas.width = smallWidth;
gl.canvas.height = smallHeight;
render();
checkForDot(gl); // fails (almost all the time)

finishTest();

var successfullyParsed = true;
</script>
</body>
</html>
